Make State trait more generic
